Revise cmd/apicompat output format and exit codes.
Remove extra newline between changes, the apicompat.Change already
formats each change with a trailing newline, and whether that's
correct or not, it's simpler the leave that behaviour as is.

The exit codes were inconsistent as well, now, if there's an internal
error the exit status will 1, if there's a breaking change it will
be 2, else it will be 0.

Fixes #34.