DeclChecker to also return the position of the change.
Previously, we were tracking the position of the general declaration
which would note the position of the start or end of the declaration.
For example:

type StructExp struct {
    MemberExp  int
    memberPriv int
}

If memberPriv was removed, the position of the struct was recorded
not the line where memberPriv was on.

Further, for some declarations, it was recording the end position
not the start position, this was intentional to work around the
fact sometimes we didn't have the start position.

This change makes it the DeclChecker's responsibility to not only
report the change, but also to return the position the change
occurred on.

In fixing this, the testdata/exp.txt results have shifted from
reporting the end position to the start, for the elements that were
affected by the end position reporting issue.